backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.2
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
requests==2.31.0
lxml==4.9.3
signxml==2.9.0
cryptography==41.0.7
reportlab==4.0.0
gunicorn==22.0.0
django-celery-beat==2.6.0
drf-nested-routers==0.9.0